Fix move package Lotto, Rank, WinningLotto Class
domain -> lotto
