+++ src/main/java/constants/LottoConstants.java
@@ -1,0 +1,6 @@
+package constants;
+
+public class LottoConstants {
+    public static final String COUNT_ERROR_MESSAGE = "로또는 6개의 숫자가 필요합니다.";
+    public static final String SAME_NUMBER_ERROR_MESSAGE = "로또는 6개의 숫자가 필요합니다.";
+}
