+++ src/main/java/lotto/Rank.java
@@ -1,4 +1,4 @@
-package domain;
+package lotto;
 
 /**
  * 로또 등수를 의미하는 enum
