+++ src/main/java/lotto/WinningLotto.java
@@ -1,4 +1,4 @@
-package domain;
+package lotto;
 
 /**
  * 당첨 번호를 담당하는 객체
